round for windows version, and one static cast to distinguish sqrt(double)

+++ src/xpcc/math/utils/pc/operator.cpp
@@ -30,10 +30,12 @@
  */
 // ----------------------------------------------------------------------------
 
-#include "operator.hpp"
+#include "../operator.hpp"
+#include <xpcc/utils/macros.hpp>
+
 
 extern "C" uint16_t
 xpcc::math::sqrt32(uint32_t input)
 {
-	return round(std::sqrt(input));
+	return round(std::sqrt(static_cast<double>(input)));
 }
